Create `react-native` entry point (#1)
