+++ apps/test-app/metro.config.js
@@ -16,4 +16,6 @@ config.resolver.nodeModulesPaths = [
   path.resolve(monorepoRoot, "node_modules"),
 ];
 
+config.resolver.unstable_enablePackageExports = true;
+
 module.exports = config;
