+++ apps/test-app/tsconfig.json
@@ -7,6 +7,7 @@
       "./node_modules/@types",
       "../../node_modules/@types",
       "../../node_modules/@webgpu/types"
-    ]
+    ],
+    "moduleResolution": "bundler"
   }
 }
