+++ packages/typegpu-confetti/package.json
@@ -1,10 +1,13 @@
 {
   "name": "typegpu-confetti",
   "version": "0.0.0",
-  "description": "Customizable React Native confetti component, running on GPU",
+  "description": "Customizable confetti animation component, running on the GPU",
   "license": "MIT",
   "type": "module",
-  "exports": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./react-native": "./src/react-native/index.ts"
+  },
   "main": "./src/index.ts",
   "types": "./src/index.ts",
   "publishConfig": {
@@ -19,6 +22,12 @@
         "module": "./dist/index.js",
         "import": "./dist/index.js",
         "default": "./dist/index.cjs"
+      },
+      "./react-native": {
+        "types": "./dist/react-native/index.d.ts",
+        "module": "./dist/react-native/index.js",
+        "import": "./dist/react-native/index.js",
+        "default": "./dist/react-native/index.cjs"
       }
     }
   },
@@ -33,7 +42,7 @@
     "type": "git",
     "url": "git+https://github.com/mhawryluk/typegpu-confetti.git"
   },
-  "keywords": ["confetti", "javascript", "react-native"],
+  "keywords": ["confetti", "javascript", "react-native", "webgpu"],
   "bugs": {
     "url": "https://github.com/mhawryluk/typegpu-confetti/issues"
   },
@@ -41,7 +50,7 @@
     "react": "18.3.1",
     "react-native": "0.76.7",
     "react-native-wgpu": "^0.1.23",
-    "typegpu": "0.4.4"
+    "typegpu": "0.5.3"
   },
   "devDependencies": {
     "typescript": "^5.7.3",
