+++ packages/typegpu-confetti/src/index.ts
@@ -1,6 +1,3 @@
-import Confetti from './Confetti';
-export { ConfettiPropTypes, ConfettiRef } from './Confetti';
-
 export {
   // slots
   canvasAspectRatio,
@@ -16,9 +13,4 @@ export {
   initParticleFn,
 } from './confetti-schemas';
 
-export {
-  ConfettiProvider,
-  useConfetti,
-} from './ConfettiProvider';
-
-export default Confetti;
+export type { ConfettiPropTypes, ConfettiRef } from './types';
