+++ packages/typegpu-confetti/src/react-native/index.ts
@@ -1,0 +1,8 @@
+import Confetti from './Confetti';
+
+export {
+  ConfettiProvider,
+  useConfetti,
+} from './ConfettiProvider';
+
+export default Confetti;
