+++ packages/typegpu-confetti/src/types.ts
@@ -1,0 +1,21 @@
+import type { TgpuFn } from 'typegpu';
+import type * as d from 'typegpu/data';
+
+export type ConfettiPropTypes = {
+  colorPalette?: [number, number, number, number][];
+  size?: number;
+  maxDurationTime?: number | null;
+
+  initParticleAmount?: number;
+  maxParticleAmount?: number;
+
+  gravity?: TgpuFn<[d.Vec2f], d.Vec2f>;
+  initParticle?: TgpuFn<[d.I32], undefined>;
+};
+
+export type ConfettiRef = {
+  pause: () => void;
+  resume: () => void;
+  restart: () => void;
+  addParticles: (amount: number) => void;
+};
