+++ pnpm-workspace.yaml
@@ -1,3 +1,6 @@
 packages:
-  - "packages/*"
-  - "apps/*"
+  - packages/*
+  - apps/*
+onlyBuiltDependencies:
+  - '@biomejs/biome'
+  - esbuild
